desktop: accounts: add port + rel field, plus shutdown fixes
Communicate on ports other than port 80.  Add a field to remote host info
that specifies which port to send http requests to.

Specify how the account state link (if any) should be shown, such as a
plain link or as a media embed.

Listen to the session manager and perform shutdown fixes.  Untested until
I can push this change to prod.  The existing method of sending
deactivation packets on exit didn't work.

Changelog:
- split away account info from host info
- add port for the host
- add rel field specifying type of account state
- work new fields into the xml saver/reader
- anything else