server: accept new media types
Accept new media types, which are communicated to the server via the
'rel' field.  'rel' can be 'text', 'href', 'audio', or 'video'.  These
will be transformed to the appropriate html tag.

Note that the api is subtly different: it accepts 'href' as a link type,
which the newer desktop client sends to indicate a html hyperlink.  With
the new changes, 'link' became ambiguous.  If it receives a 'link' rel
type, it is mangled to 'href' to accomodate the whims of old clients.
Please update your clients.

Because media files are now embedded, your streaming servers should send
the http header `Cache-Control: no-cache` so the browser doesn't cache
old streams.